ご指摘いただいたディレクトリ構成にリファクタリング
