--- components/model/sidebar/index.ts
@@ -1,1 +1,0 @@
-export * from './SideBar';
