--- components/page/index.ts
@@ -1,2 +1,0 @@
-export * from './Main.page';
-export * from './components/MainContents';
