+++ components/shared/contents/index.ts
@@ -1,0 +1,2 @@
+export * from './Contents';
+export * from './ContentsHeader';
