+++ components/shared/sidebar/index.ts
@@ -1,0 +1,3 @@
+export * from './SideBar';
+export * from './SideBarHeader';
+export * from './SideBarContents';
